==> jupyter_notebooks/02 - Patients Study.ipynb <==
# Churned Customer Study Notebook

## Objectives

*   Answer business requirement 1: 
    * The client is interested to understand the patterns from the patient data in order to better predict the likelyhood of heart failure in a patient.

## Inputs

* outputs/datasets/collection/heart-failure.csv

## Outputs

* generate code that answers business requirement 1 and can be used to build Streamlit App






---

# Change working directory

We need to change the working directory from its current folder to its parent folder
* We access the current directory with os.getcwd()

import os
current_dir = os.getcwd()
current_dir

We want to make the parent of current directory the new current directory
* os.path.dirname() gets the parent directory
* os.chir() defines the new current directory

os.chdir(os.path.dirname(current_dir))
print("You set a new current directory")

Confirm the new current directory

current_dir = os.getcwd()
current_dir

---

# Load Data

import pandas as pd
df = (pd.read_csv("outputs/datasets/collection/heart-failure.csv"))
df.head(3)

# Data Exploration

We are interested to get more familiar with the dataset, check variable type and distribution, missing levels and what these variables mean in a business context

from pandas_profiling import ProfileReport
pandas_report = ProfileReport(df=df, minimal=True)
pandas_report.to_notebook_iframe()

# Correlation Study

We use `.corr()` for `spearman` and `pearson` methods, and investigate the top 10 correlations
* We know this command returns a pandas series and the first item is the correlation between Churn and Churn, which happens to be 1, so we exclude that with `[1:]`
* We sort values considering the aboslute value, by setting `key=abs`

corr_spearman = df.corr(method='spearman')['DEATH_EVENT'].sort_values(key=abs, ascending=False)[1:].head(10)
corr_spearman

We do the same for `pearson`

corr_pearson = df.corr(method='pearson')['DEATH_EVENT'].sort_values(key=abs, ascending=False)[1:].head(10)
corr_pearson

For both methods, we notice weak or moderate levels of correlation between DEATH_EVENT and a given variable. 
* Ideally, we pursue at strong correlation levels or more. However, this is not always possible

We will consider the top 7 correlation levels at `df`

top_n = 5
set(corr_pearson[:top_n].index.to_list() + corr_spearman[:top_n].index.to_list())

Therefore we are studying at df the following variables. We will investigate if:
* A patient who dies of heart failure typically has low/high age (to be edited)
* A patient who dies of heart failure typically has low/high ejection fraction (to be edited)
* A patient who dies of heart failure typically has low/high Level of serum creatinine in the blood (mg/dL) (to be edited)
* A patient who dies of heart failure typically has low/high Level of serum sodium in the blood (mEq/L) (to be edited)
* A patient who dies of heart failure typically has low/high Follow-up period (days) (to be edited)

vars_to_study = ['age', 'ejection_fraction', 'serum_creatinine', 'serum_sodium', 'time']
vars_to_study

# EDA on selected variables

df_eda = df.filter(vars_to_study + ['DEATH_EVENT'])
df_eda.head(3)

## Variables Distribution by Death Event

We plot the distribution (numerical and categorical) colored by Death Event

import matplotlib.pyplot as plt
import seaborn as sns
sns.set_style('whitegrid')

def plot_categorical(df, col, target_var):

  plt.figure(figsize=(12, 5))
  sns.countplot(data=df, x=col, hue=target_var,order = df[col].value_counts().index)
  plt.xticks(rotation=90) 
  plt.title(f"{col}", fontsize=20,y=1.05)        
  plt.show()

def plot_numerical(df, col, target_var):
  plt.figure(figsize=(8, 5))
  sns.histplot(data=df, x=col, hue=target_var, kde=True,element="step") 
  plt.title(f"{col}", fontsize=20,y=1.05)
  plt.show()



target_var = 'DEATH_EVENT'
for col in vars_to_study:
  if df_eda[col].dtype == 'object':
    plot_categorical(df_eda, col, target_var)
    print("\n\n")
  else:
    plot_numerical(df_eda, col, target_var)
    print("\n\n")

---

## Parallel Plot

Create a separate DataFrame and transform `time` (numerical) into bins (categorical) for visualizing at `parallel_categories()` plot


from feature_engine.discretisation import ArbitraryDiscretiser
import numpy as np
time_map = [-np.Inf, 50, 100, 150, 200, np.Inf]
sodium_map = [-np.Inf, 135, np.inf]
creatinine_map = [-np.Inf, 1, np.inf]
ejection_map = [-np.Inf, 30, 40, np.inf]
age_map = [-np.Inf, 50, 60, 70, np.inf]
maps = [age_map, sodium_map, creatinine_map, ejection_map, time_map]
variables = ['age', 'serum_sodium', 'serum_creatinine', 'ejection_fraction', 'time']
disc = ArbitraryDiscretiser(binning_dict={
    'age': age_map,
    'serum_sodium': sodium_map,
    'serum_creatinine': creatinine_map,
    'ejection_fraction': ejection_map,
    'time': time_map})
df_parallel = disc.fit_transform(df_eda)
df_parallel.head()

Create map to replace variable labels with more informative levels

labels_dicts = {}
for i in range(0, len(variables)):
  n_classes = len(maps[i]) - 1
  classes_ranges = disc.binner_dict_[variables[i]][1:-1]

  labels_map = {}
  for n in range(0,n_classes):
    if n == 0:
      labels_map[n] = f"<{classes_ranges[0]}"
    elif n == n_classes-1:
      labels_map[n] = f"+{classes_ranges[-1]}"
    else:
      labels_map[n] = f"{classes_ranges[n-1]} to {classes_ranges[n]}"

  labels_dicts[variables[i]] = labels_map
labels_dicts

Replace using `.replace()`

for variable in variables:
    df_parallel[variable] = df_parallel[variable].replace(
        labels_dicts[variable])
df_parallel.head()

Creates multi-dimensional categorical data plot

import plotly.express as px
fig = px.parallel_categories(df_parallel, color="DEATH_EVENT")
fig.show()

---

# Conclusions and Next steps

Patients are more at risk of dying of heart failure if they:
* Are older
* Have a low ejection fraction (the percentage of blood moved on each heartbeat)
* Have a creatinine level higher than 1 in their serum
* Have a sodium level lower than 135 in their serum
* Have a short follow-up period (typically, the longer a patient stays alive while being followed up, the lower the odds of a death event still happening)
